+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - When the user searches for a specific set of games, the `add all` button is replaced by the `add all results` button
   - Unlike `add all`, the `add all results` button will append the results to the list, rather than overwriting the existing list
   - This allows users to search for multiple different groups of games and build a custom list easily
+- Added a `recommended` carousel to `card farming`. This will recommend 10 randomly picked games from your library that have card drops remaining
+  - You can easily add recommended games to your list by clicking the "+" button below the cover art
 - Fixed an issue where the Close button in Card Farming screen would stop the scheduled Automatic Idler task (#770)
 - In the portable version, achievement data files are now correctly created and stored in the same directory as the executable